R code cleanup -> R CMD Check passes (notes only)
* Generics used from other packages must be reexported
* Some missing args now documented
* Unpleasantness with S3 methods (see below)

Previously, all S3 methods for `epi_df` had `x` as the first argument. But this doesn't actually pass checks. Any method must match the generic signature. So this means replacing `x` with (variously) `object`/`.data`/`data`. This is a bit annoying to document all together. Adding an issue to split the methods up in the docs.