Refactor `ref_lag` from global to params, get tests working
Get lag-completion/target-lag function tests working in package setting. Before
this change, `ref_lag` was a global set in a not-yet-package file and also the
test file, which, when converted into a package, did not pass tests, as the
tests' `ref_lag` global doesn't overwrite the pre-existing value from the
package.