Emit some help on `archive_cases_dv` eval errors during `unregister`
Package reloading can/will force package-level promises, so if we introduce a
bug in package-level promises, we can get "trapped" and unable to reload without
using `unloadNamespace` or more drastic measures like restarting the R session.
But we may be confused at what's going on in the first place when this happens,
so build the promise to emit a helpful message if it looks like we're in that
situation.