Go back to using NA to indicate no clobberable versions
While drafting, NA was used originally, but this ran the danger of the user
using `max` on a version column with 0 entries or an `NA`.  However, NA versions
are disallowed now, so only the 0-entry case is to be worried about, and this is
only if the user uses `max` themselves instead of relying on our checked default
or the function we use to prepare this default.

`NULL` runs a much greater danger: it appears that R6 does not add any extra
checks to field/method access to make sure that things actually exist, and, like
base environments, returns NULL in the case of certain typos or an object that
was invalidly constructed without the expected field.

A third option, a custom sentinel value, would be the most robust, but would
also pollute the namespace.  Try to avoid this for now.