Fix example archive to match old val
Use LVCF, not NA, to fill in versions in the less up-to-date archive. This
should get closer to matching the old data, but still needs to be checked. The
old nafill would have filled in some things we wouldn't want, but this doesn't
seem to be one of these scenarios, as git thinks the generated sysdata is the
same.