Make epix_slide tidyselect tests tougher (not just default grping)
Existing tests of the resolution of some rlang&tidyselect issues in epix_slide
were a bit loose, testing only different ways of explicitly specifying a
grouping that all were equal to the default grouping. This came into play with a
test that contained faulty usage of `!!!` which appeared to be somehow
transformed by testthat expectations into something zapping the `group_by` arg
from the related call. This tougher line of testing is intended to catch any
similar types of issues with the other test cases.