Avoid `filter(cases==1)`, better describe & `complete` Ebola example
Avoid `filter(cases==1)`: this should either

* involve `filter(status == "confirmed")` instead, or
* involve no filter, and let `sum` take care of things.

We should generally expect the latter approach to be RAM-friendlier (one new
column vs. new filtered vectors for every column), so implement the latter
approach.

Reformat the code a bit to try to make it read better.

Clarify the description of the example; make clear we're deriving from line list
data.

`complete` the data to make the set of `time_value`s evenly spaced and the same
for every `geo_value`; even though it doesn't impact the plots, we're
demonstrating how to get things into `epi_df` format, and many functions we'd
use on `epi_df`s will expect things in this format.