turn epix_truncate_versions_after into a generic
When `x` is a grouped `epi_archive`, it must be cloned and ungrouped so
basic checks on `max_version` can be done. The cloning step is expensive
and is unnecessary when the call is invalid or when the archive is not
grouped. Separate the fn calls to simplify logic and avoid the
sometimes-unecessary `clone`.