deduplicate function/formula and quosure slides
In `as_slide_computation`, call `eval_tidy` on the quosure `x` passed
directly to `as_slide_computation`, rather than generating an
`f_wrapper` computation function that takes the quosure as an argument.
The computation function is regenerated each time `slide` is called, with
a new quosure, so the computation function doesn't need to be flexible
enough to run with different `quo`s.

This change makes the function/formula and quosure forks more similar,
since `group_modify`, `slide_one_grp` in the `epi_slide` case, and
`comp_one_grp` in the `epix_slide` case no longer need a `quo` argument
in the quosure fork. To make the two forks fully identical, the quosure
fork was changed to pass an empty set of dots to the computation
functions.

The `as_slide_computation` call and `group_modify` call can now be
pulled out of the if/else block.