fix(epi[x]_slide): again, don't reject (row)named data frames
Fix another place where these were being rejected. Not making a helper function
yet because this particular instance of the check is in a tight loop and a
function call (maybe multiple if we want have a vectorized check) might be
substantial overhead.  Want to check for performance regressions in existing
changes before trying this type of refactor.