Remove unnecessary/misleading `select.epi_df` implementation
`select.epi_df` looked weird:
- `reclass(selected, attr(selected, "metadata"))` would only have made sense if
  grouped_df had a `select` that dropped our class but not our metadata,
  implemented in a way that called our own impls for `dplyr_extending` that would
  make that metadata actually be correct (in the case of renaming).
- (`dplyr_reconstruct(selected, selected)` might have made sense, if `reclass`
  acted on something that shouldn't actually become an `epi_df`, decaying back to
  a non-`epi_df` only if needed.)

So it seemed like it relied on `dplyr_extending` being almost, but not quite,
sufficient. But it looks like `grouped_df` doesn't even have a `select` impl, so
it's not interfering anyway! So it looks like we can just get rid of our own
impl.