Fix empty-version grouped slide behavior change, tweak warning
Moving from `group_modify` to `group_map` means doing our own labeling. Handle a
corner case here in a way similar to `group_modify` (if you think of `version`
being a non-dropping factor at the beginning of the grouping columns). It's not
clear this behavior is what we want... perhaps we want to generate errors if we
ever have NAs generated from this not-a-nondropping-factor
version/grouping-column following a nondropping-factor version/grouping-column.
Or at least warn in the subcase of this when a user requests an empty version
when `.drop = TRUE` or `.drop = FALSE` but there are nonfactor columns.

Tweak the `.drop = FALSE` warning that tries to ward against some of the above
cases (but is only thinking about grouping columns rather than version +
grouping columns) so that it won't warn if there are 0 grouping columns. It
makes sense that a user may have a set of potential factor grouping columns and
programmatically try models with different grouping column sets and `.drop =
FALSE`, including an empty grouping column set.