docs(epi_df.Rmd): immediate ungrouping + discuss completion effects
It's probably best to immediately ungroup after performing grouped operations in
our documentation, as leaving things grouped accidentally is a source of errors.
Sometime we should consider an overhaul to use `by =` and `.by =` where
appropriate (sorting effects not needed) and available (not all operations
support this syntax yet).

There were already 0s in the example data, so "highlight" with words the effects
of completion + note one potential surprise in other applications.